continued modularising the code base
